Cleanup

+++ render/include/rawrbox/render/text/engine.hpp
@@ -12,18 +12,18 @@ namespace rawrbox {
 	class TextEngine {
 
 		static std::map<std::string, std::unique_ptr<rawrbox::Font>> _fonts;
-		static std::map<uint32_t, std::unique_ptr<rawrbox::TexturePack>> _packs;
+		static std::map<uint16_t, std::unique_ptr<rawrbox::TexturePack>> _packs;
 
 		static std::string getFontInSystem(const std::filesystem::path& path);
 
 	public:
-		static uint32_t packID;
+		static uint16_t packID;
 
 		static void shutdown();
 
-		static std::pair<uint32_t, rawrbox::TexturePack*> requestPack(int width, int height, bgfx::TextureFormat::Enum format = bgfx::TextureFormat::RGBA8);
-		static rawrbox::TexturePack* getPack(uint32_t id);
+		static std::pair<uint16_t, rawrbox::TexturePack*> requestPack(uint16_t width, uint16_t height, bgfx::TextureFormat::Enum format = bgfx::TextureFormat::RGBA8);
+		static rawrbox::TexturePack* getPack(uint16_t id);
 
-		static rawrbox::Font* load(const std::filesystem::path& filename, uint32_t size, uint32_t index = 0);
+		static rawrbox::Font* load(const std::filesystem::path& filename, uint16_t size, uint16_t index = 0);
 	};
 } // namespace rawrbox
